drm/rockchip: dw_hdmi: make vop source selection part of platform_data
The vop source for the dw_hdmi on Rockchip socs is done in the General
Register Files and thus will likely vary on each new soc using it.
Therefore introduce a separate platform_data struct for the Rockchip
variants of dw_hdmi and put the register and setting bits into it.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>